# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "X API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/posts": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "maxLength": 280
                  },
                  "mediaIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^\\d+$"
                    },
                    "maxItems": 4
                  },
                  "inReplyToId": {
                    "type": "string",
                    "pattern": "^\\d+$"
                  },
                  "quotedPostId": {
                    "type": "string",
                    "pattern": "^\\d+$"
                  },
                  "replyPolicy": {
                    "type": "string",
                    "enum": [
                      "everyone",
                      "following",
                      "mentioned"
                    ],
                    "default": "everyone"
                  },
                  "isSensitive": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": false
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "pattern": "^\\d+$"
                        },
                        "text": {
                          "type": "string",
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "author": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "pattern": "^\\d+$"
                            },
                            "username": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 15,
                              "pattern": "^[A-Za-z0-9_]+$"
                            },
                            "displayName": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 50
                            },
                            "avatarUrl": {
                              "type": "string",
                              "format": "uri",
                              "nullable": true
                            },
                            "isVerified": {
                              "type": "boolean"
                            }
                          },
                          "required": [
                            "id",
                            "username",
                            "displayName",
                            "avatarUrl",
                            "isVerified"
                          ],
                          "additionalProperties": false
                        },
                        "entities": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "kind": {
                                "type": "string",
                                "enum": [
                                  "mention",
                                  "hashtag",
                                  "url",
                                  "cashtag"
                                ]
                              },
                              "value": {
                                "type": "string"
                              },
                              "start": {
                                "type": "integer",
                                "minimum": 0
                              },
                              "end": {
                                "type": "integer",
                                "minimum": 0
                              }
                            },
                            "required": [
                              "kind",
                              "value",
                              "start",
                              "end"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "media": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "pattern": "^\\d+$"
                              },
                              "kind": {
                                "type": "string",
                                "enum": [
                                  "image",
                                  "gif",
                                  "video"
                                ]
                              },
                              "url": {
                                "type": "string",
                                "format": "uri"
                              },
                              "width": {
                                "type": "integer",
                                "exclusiveMinimum": true,
                                "minimum": 0,
                                "nullable": true
                              },
                              "height": {
                                "type": "integer",
                                "exclusiveMinimum": true,
                                "minimum": 0,
                                "nullable": true
                              },
                              "blurhash": {
                                "type": "string",
                                "nullable": true
                              },
                              "altText": {
                                "type": "string",
                                "nullable": true
                              }
                            },
                            "required": [
                              "id",
                              "kind",
                              "url",
                              "width",
                              "height",
                              "blurhash",
                              "altText"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "conversationId": {
                          "type": "string",
                          "pattern": "^\\d+$"
                        },
                        "inReplyToId": {
                          "type": "string",
                          "pattern": "^\\d+$",
                          "nullable": true
                        },
                        "counters": {
                          "type": "object",
                          "properties": {
                            "likes": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "reposts": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "replies": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "quotes": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "views": {
                              "type": "integer",
                              "minimum": 0
                            }
                          },
                          "required": [
                            "likes",
                            "reposts",
                            "replies",
                            "quotes",
                            "views"
                          ],
                          "additionalProperties": false
                        },
                        "viewer": {
                          "type": "object",
                          "properties": {
                            "liked": {
                              "type": "boolean"
                            },
                            "reposted": {
                              "type": "boolean"
                            },
                            "bookmarked": {
                              "type": "boolean"
                            }
                          },
                          "required": [
                            "liked",
                            "reposted",
                            "bookmarked"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "id",
                        "text",
                        "createdAt",
                        "author",
                        "entities",
                        "media",
                        "conversationId",
                        "inReplyToId",
                        "counters"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "BLOCKED_BY_USER",
                            "NOT_FOUND",
                            "CONFLICT",
                            "UNPROCESSABLE",
                            "RATE_LIMIT_EXCEEDED",
                            "INTERNAL_ERROR",
                            "SERVICE_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "BLOCKED_BY_USER",
                            "NOT_FOUND",
                            "CONFLICT",
                            "UNPROCESSABLE",
                            "RATE_LIMIT_EXCEEDED",
                            "INTERNAL_ERROR",
                            "SERVICE_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "BLOCKED_BY_USER",
                            "NOT_FOUND",
                            "CONFLICT",
                            "UNPROCESSABLE",
                            "RATE_LIMIT_EXCEEDED",
                            "INTERNAL_ERROR",
                            "SERVICE_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/v1"
    }
  ]
}